Refactor _common folder approach
We have a lot code duplication as well as potential code dup
that will be written in future. It's needed to refactor
the approach we have for the common folder.

In this commit was created new static public library "common"
main aim of what will be to avoid violation of DRY principle.

To make this global refactoring also was needed to do changes:
- ApiManager and all network related logic will use std::string
- Serializers are moved to common

Signed-off-by: Sage-Cat <sagecatbox@gmail.com>