Refactor MainWindow arch. Implement tabWidget
We have overkilling arch with UiManager. It's
enough to have MainWindow as "UiManager" that will
do all the work. There is no good reason for creating
UiManager class for such a work. Same for SageStoreClient

Also we have no implementation for TabWidget. We need it
to be able to implement further features

Signed-off-by: Sage-Cat <sagecatbox@gmail.com>